+++ src/Passes/Source/AllocationManager/IAllocationManager.hpp
@@ -9,23 +9,21 @@
 #include <unordered_map>
 #include <vector>
 
-namespace microsoft
+namespace microsoft {
+namespace quantum {
+
+class IAllocationManager
 {
-namespace quantum
-{
+public:
+  using Address              = uint64_t;
+  using Index                = uint64_t;
+  using String               = std::string;
+  using AllocationManagerPtr = std::shared_ptr<IAllocationManager>;
 
-    class IAllocationManager
-    {
-      public:
-        using Address              = uint64_t;
-        using Index                = uint64_t;
-        using String               = std::string;
-        using AllocationManagerPtr = std::shared_ptr<IAllocationManager>;
+  virtual ~IAllocationManager();
+  virtual Address allocate(String const &name = "", Index const &size = 1) = 0;
+  virtual void    release(Address const &address)                          = 0;
+};
 
-        virtual ~IAllocationManager();
-        virtual Address allocate(String const& name = "", Index const& size = 1) = 0;
-        virtual void    release(Address const& address)                          = 0;
-    };
-
-} // namespace quantum
-} // namespace microsoft
+}  // namespace quantum
+}  // namespace microsoft
