+++ src/Passes/Source/Apps/CMakeLists.txt
@@ -1,4 +1,4 @@
-add_executable(qat Qat/Qat.cpp Qat/LlvmAnalysis.cpp Qat/Config.cpp)
+add_executable(qat Qat/Qat.cpp Qat/Config.cpp)
 
 target_link_libraries(qat ${llvm_libs})
-target_link_libraries(qat RuleTransformationPass Rules AllocationManager Commandline Generators)
+target_link_libraries(qat RuleTransformationPass Rules AllocationManager Commandline Generators Profile)
