+++ src/Passes/Source/Apps/Qat/Qat.cpp
@@ -43,15 +43,14 @@
 ///
 
 #include "Apps/Qat/Config.hpp"
-#include "Apps/Qat/LlvmAnalysis.hpp"
 #include "Commandline/ConfigurationManager.hpp"
 #include "Commandline/ParameterParser.hpp"
 #include "Generators/DefaultProfileGenerator.hpp"
 #include "Generators/LlvmPassesConfig.hpp"
+#include "Llvm/Llvm.hpp"
+#include "Profile/Profile.hpp"
 #include "RuleTransformationPass/Configuration.hpp"
 #include "Rules/FactoryConfig.hpp"
-
-#include "Llvm/Llvm.hpp"
 
 #include <iomanip>
 #include <iostream>
@@ -60,127 +59,127 @@
 using namespace llvm;
 using namespace microsoft::quantum;
 
-int main(int argc, char** argv)
+int main(int argc, char **argv)
 {
-    try
+  try
+  {
+    auto generator = std::make_shared<DefaultProfileGenerator>();
+
+    ConfigurationManager &configuration_manager = generator->configurationManager();
+    configuration_manager.addConfig<QatConfig>();
+    configuration_manager.addConfig<FactoryConfiguration>();
+
+    // Parsing command line arguments
+    ParameterParser parser;
+    configuration_manager.setupArguments(parser);
+    parser.parseArgs(argc, argv);
+    configuration_manager.configure(parser);
+
+    // Getting the main configuration
+    auto const &config = configuration_manager.get<QatConfig>();
+
+    // In case we debug, we also print the settings to allow provide a full
+    // picture of what is going.
+    if (config.dumpConfig())
     {
-        auto profile = std::make_shared<DefaultProfileGenerator>();
-
-        ConfigurationManager& configuration_manager = profile->configurationManager();
-        configuration_manager.addConfig<QatConfig>();
-        configuration_manager.addConfig<FactoryConfiguration>();
-
-        // Parsing command line arguments
-        ParameterParser parser;
-        configuration_manager.setupArguments(parser);
-        parser.parseArgs(argc, argv);
-        configuration_manager.configure(parser);
-
-        // Getting the main configuration
-        auto const& config = configuration_manager.get<QatConfig>();
-
-        // In case we debug, we also print the settings to allow provide a full
-        // picture of what is going.
-        if (config.dumpConfig())
-        {
-            configuration_manager.printConfiguration();
-        }
-
-        if (parser.arguments().empty())
-        {
-            std::cerr << "Usage: " << argv[0] << " [options] filename" << std::endl;
-            configuration_manager.printHelp();
-            std::cerr << "\n";
-            exit(-1);
-        }
-
-        // Loading IR from file.
-        LLVMContext  context;
-        SMDiagnostic error;
-        auto         module = parseIRFile(parser.getArg(0), error, context);
-
-        if (!module)
-        {
-            std::cerr << "Invalid IR." << std::endl;
-            exit(-1);
-        }
-
-        // Extracting commandline parameters
-
-        auto optimisation_level = llvm::PassBuilder::OptimizationLevel::O0;
-
-        // Setting the optimisation level
-        if (config.opt1())
-        {
-            optimisation_level = llvm::PassBuilder::OptimizationLevel::O1;
-        }
-
-        if (config.opt2())
-        {
-            optimisation_level = llvm::PassBuilder::OptimizationLevel::O2;
-        }
-
-        if (config.opt3())
-        {
-            optimisation_level = llvm::PassBuilder::OptimizationLevel::O3;
-        }
-
-        // Checking if we are asked to generate a new QIR. If so, we will use
-        // the profile to setup passes to
-        if (config.generate())
-        {
-            // Creating pass builder
-            LlvmAnalyser analyser{config.debug()};
-
-            // Preparing pass for generation based on profile
-            profile->addFunctionAnalyses(analyser.functionAnalysisManager());
-            auto module_pass_manager =
-                profile->createGenerationModulePass(analyser.passBuilder(), optimisation_level, config.debug());
-
-            // Running the pass built by the profile
-            module_pass_manager.run(*module, analyser.moduleAnalysisManager());
-
-            // Priniting either human readible LL code or byte
-            // code as a result, depending on the users preference.
-            if (config.emitLlvm())
-            {
-                llvm::errs() << *module << "\n";
-            }
-            else
-            {
-                llvm::errs() << "Byte code ouput is not supported yet. Please add -S to get human readible "
-                                "LL code.\n";
-            }
-
-            // Verifying the module.
-            if (config.verifyModule())
-            {
-                llvm::VerifierAnalysis verifier;
-                auto                   result = verifier.run(*module, analyser.moduleAnalysisManager());
-                if (result.IRBroken)
-                {
-                    llvm::errs() << "IR is broken."
-                                 << "\n";
-                    exit(-1);
-                }
-            }
-        }
-
-        if (config.validate())
-        {
-            // Creating pass builder
-            LlvmAnalyser analyser{config.debug()};
-
-            // Creating a validation pass manager
-            auto module_pass_manager =
-                profile->createValidationModulePass(analyser.passBuilder(), optimisation_level, config.debug());
-            module_pass_manager.run(*module, analyser.moduleAnalysisManager());
-        }
-    }
-    catch (std::exception const& e)
-    {
-        llvm::errs() << "An error occured: " << e.what() << "\n";
+      configuration_manager.printConfiguration();
     }
 
-    return 0;
+    if (parser.arguments().empty())
+    {
+      std::cerr << "Usage: " << argv[0] << " [options] filename" << std::endl;
+      configuration_manager.printHelp();
+      std::cerr << "\n";
+      exit(-1);
+    }
+
+    // Loading IR from file.
+    LLVMContext  context;
+    SMDiagnostic error;
+    auto         module = parseIRFile(parser.getArg(0), error, context);
+
+    if (!module)
+    {
+      std::cerr << "Invalid IR." << std::endl;
+      exit(-1);
+    }
+
+    // Extracting commandline parameters
+
+    auto optimisation_level = llvm::PassBuilder::OptimizationLevel::O0;
+
+    // Setting the optimisation level
+    if (config.opt1())
+    {
+      optimisation_level = llvm::PassBuilder::OptimizationLevel::O1;
+    }
+
+    if (config.opt2())
+    {
+      optimisation_level = llvm::PassBuilder::OptimizationLevel::O2;
+    }
+
+    if (config.opt3())
+    {
+      optimisation_level = llvm::PassBuilder::OptimizationLevel::O3;
+    }
+
+    // Checking if we are asked to generate a new QIR. If so, we will use
+    // the profile to setup passes to
+    if (config.generate())
+    {
+      // Creating pass builder
+      Profile analyser{config.debug()};
+
+      // Preparing pass for generation based on profile
+      generator->addFunctionAnalyses(analyser.functionAnalysisManager());
+      auto module_pass_manager = generator->createGenerationModulePass(
+          analyser.passBuilder(), optimisation_level, config.debug());
+
+      // Running the pass built by the profile
+      module_pass_manager.run(*module, analyser.moduleAnalysisManager());
+
+      // Priniting either human readible LL code or byte
+      // code as a result, depending on the users preference.
+      if (config.emitLlvm())
+      {
+        llvm::errs() << *module << "\n";
+      }
+      else
+      {
+        llvm::errs() << "Byte code ouput is not supported yet. Please add -S to get human readible "
+                        "LL code.\n";
+      }
+
+      // Verifying the module.
+      if (config.verifyModule())
+      {
+        llvm::VerifierAnalysis verifier;
+        auto                   result = verifier.run(*module, analyser.moduleAnalysisManager());
+        if (result.IRBroken)
+        {
+          llvm::errs() << "IR is broken."
+                       << "\n";
+          exit(-1);
+        }
+      }
+    }
+
+    if (config.validate())
+    {
+      // Creating pass builder
+      Profile analyser{config.debug()};
+
+      // Creating a validation pass manager
+      auto module_pass_manager = generator->createValidationModulePass(
+          analyser.passBuilder(), optimisation_level, config.debug());
+      module_pass_manager.run(*module, analyser.moduleAnalysisManager());
+    }
+  }
+  catch (std::exception const &e)
+  {
+    llvm::errs() << "An error occured: " << e.what() << "\n";
+  }
+
+  return 0;
 }
