+++ src/Passes/Source/CMakeLists.txt
@@ -6,6 +6,9 @@ microsoft_add_library_tests(TestTools)
 microsoft_add_library(Commandline)
 microsoft_add_library_tests(Commandline)
 
+microsoft_add_library(Profile)
+microsoft_add_library_tests(Profile)
+
 microsoft_add_library(AllocationManager)
 microsoft_add_library_tests(AllocationManager)
 
@@ -20,7 +23,7 @@ microsoft_add_library_tests(Rules Rules RuleTransformationPass AllocationManager
 target_link_libraries(Rules PRIVATE AllocationManager)
 
 microsoft_add_library(Generators)
-target_link_libraries(Generators PRIVATE Rules Commandline AllocationManager)
+target_link_libraries(Generators PRIVATE Rules Commandline AllocationManager Profile)
 microsoft_add_library_tests(Generators Rules RuleTransformationPass AllocationManager Commandline Generators)
 
 add_subdirectory(Apps)
