+++ src/Passes/Source/Generators/DefaultProfileGenerator.cpp
@@ -2,102 +2,101 @@
 // Licensed under the MIT License.
 
 #include "Generators/DefaultProfileGenerator.hpp"
-#include "RuleTransformationPass/Profile.hpp"
+
+#include "Llvm/Llvm.hpp"
+#include "RuleTransformationPass/RulePass.hpp"
 #include "Rules/Factory.hpp"
 #include "Rules/FactoryConfig.hpp"
 #include "Rules/RuleSet.hpp"
 
-#include "Llvm/Llvm.hpp"
-
 #include <iostream>
 
-namespace microsoft
+namespace microsoft {
+namespace quantum {
+
+DefaultProfileGenerator::DefaultProfileGenerator()
 {
-namespace quantum
+  registerProfileComponent<RuleTransformationPassConfiguration>(
+      "transformation-rules",
+      [this](RuleTransformationPassConfiguration const &config, IProfileGenerator *ptr) {
+        auto &ret = ptr->modulePassManager();
+
+        // Defining the mapping
+        RuleSet rule_set;
+        auto    factory = RuleFactory(rule_set);
+        factory.usingConfiguration(configurationManager().get<FactoryConfiguration>());
+
+        // Creating profile pass
+        ret.addPass(RuleTransformationPass(std::move(rule_set), config));
+      });
+
+  registerProfileComponent<LlvmPassesConfiguration>(
+      "llvm-passes", [](LlvmPassesConfiguration const &config, IProfileGenerator *ptr) {
+        // Configuring LLVM passes
+        if (config.alwaysInline())
+        {
+          auto &ret          = ptr->modulePassManager();
+          auto &pass_builder = ptr->passBuilder();
+          ret.addPass(llvm::AlwaysInlinerPass());
+
+          auto inliner_pass = pass_builder.buildInlinerPipeline(
+              ptr->optimisationLevel(), llvm::PassBuilder::ThinLTOPhase::None, ptr->debug());
+          ret.addPass(std::move(inliner_pass));
+        }
+      });
+}
+
+DefaultProfileGenerator::DefaultProfileGenerator(
+    ConfigureFunction const &                  configure,
+    RuleTransformationPassConfiguration const &profile_pass_config,
+    LlvmPassesConfiguration const &            llvm_config)
 {
+  registerProfileComponent<RuleTransformationPassConfiguration>(
+      "Transformation rules",
+      [configure](RuleTransformationPassConfiguration const &config, IProfileGenerator *ptr) {
+        // Defining the mapping
+        auto &  ret = ptr->modulePassManager();
+        RuleSet rule_set;
+        auto    factory = RuleFactory(rule_set);
+        configure(rule_set);
 
-    DefaultProfileGenerator::DefaultProfileGenerator()
-    {
-        registerProfileComponent<RuleTransformationPassConfiguration>(
-            "transformation-rules", [this](RuleTransformationPassConfiguration const& config, IProfileGenerator* ptr) {
-                auto& ret = ptr->modulePassManager();
+        // Creating profile pass
+        ret.addPass(RuleTransformationPass(std::move(rule_set), config));
+      });
 
-                // Defining the mapping
-                RuleSet rule_set;
-                auto    factory = RuleFactory(rule_set);
-                factory.usingConfiguration(configurationManager().get<FactoryConfiguration>());
+  registerProfileComponent<LlvmPassesConfiguration>(
+      "llvm-passes", [](LlvmPassesConfiguration const &config, IProfileGenerator *ptr) {
+        // Configuring LLVM passes
+        if (config.alwaysInline())
+        {
+          auto &ret          = ptr->modulePassManager();
+          auto &pass_builder = ptr->passBuilder();
+          ret.addPass(llvm::AlwaysInlinerPass());
 
-                // Creating profile pass
-                ret.addPass(RuleTransformationPass(std::move(rule_set), config));
-            });
+          auto inliner_pass = pass_builder.buildInlinerPipeline(
+              ptr->optimisationLevel(), llvm::PassBuilder::ThinLTOPhase::None, ptr->debug());
+          ret.addPass(std::move(inliner_pass));
+        }
+      });
 
-        registerProfileComponent<LlvmPassesConfiguration>(
-            "llvm-passes", [](LlvmPassesConfiguration const& config, IProfileGenerator* ptr) {
-                // Configuring LLVM passes
-                if (config.alwaysInline())
-                {
-                    auto& ret          = ptr->modulePassManager();
-                    auto& pass_builder = ptr->passBuilder();
-                    ret.addPass(llvm::AlwaysInlinerPass());
+  configurationManager().setConfig(profile_pass_config);
+  configurationManager().setConfig(llvm_config);
+}
 
-                    auto inliner_pass = pass_builder.buildInlinerPipeline(
-                        ptr->optimisationLevel(), llvm::PassBuilder::ThinLTOPhase::None, ptr->debug());
-                    ret.addPass(std::move(inliner_pass));
-                }
-            });
-    }
+FactoryConfiguration const &DefaultProfileGenerator::factoryConfig() const
+{
+  return configurationManager().get<FactoryConfiguration>();
+}
 
-    DefaultProfileGenerator::DefaultProfileGenerator(
-        ConfigureFunction const&                   configure,
-        RuleTransformationPassConfiguration const& profile_pass_config,
-        LlvmPassesConfiguration const&             llvm_config)
-    {
-        registerProfileComponent<RuleTransformationPassConfiguration>(
-            "Transformation rules",
-            [configure](RuleTransformationPassConfiguration const& config, IProfileGenerator* ptr) {
-                // Defining the mapping
-                auto&   ret = ptr->modulePassManager();
-                RuleSet rule_set;
-                auto    factory = RuleFactory(rule_set);
-                configure(rule_set);
+RuleTransformationPassConfiguration const &DefaultProfileGenerator::profilePassConfig() const
+{
+  return configurationManager().get<RuleTransformationPassConfiguration>();
+}
 
-                // Creating profile pass
-                ret.addPass(RuleTransformationPass(std::move(rule_set), config));
-            });
+LlvmPassesConfiguration const &DefaultProfileGenerator::llvmConfig() const
+{
+  return configurationManager().get<LlvmPassesConfiguration>();
+}
 
-        registerProfileComponent<LlvmPassesConfiguration>(
-            "llvm-passes", [](LlvmPassesConfiguration const& config, IProfileGenerator* ptr) {
-                // Configuring LLVM passes
-                if (config.alwaysInline())
-                {
-                    auto& ret          = ptr->modulePassManager();
-                    auto& pass_builder = ptr->passBuilder();
-                    ret.addPass(llvm::AlwaysInlinerPass());
-
-                    auto inliner_pass = pass_builder.buildInlinerPipeline(
-                        ptr->optimisationLevel(), llvm::PassBuilder::ThinLTOPhase::None, ptr->debug());
-                    ret.addPass(std::move(inliner_pass));
-                }
-            });
-
-        configurationManager().setConfig(profile_pass_config);
-        configurationManager().setConfig(llvm_config);
-    }
-
-    FactoryConfiguration const& DefaultProfileGenerator::factoryConfig() const
-    {
-        return configurationManager().get<FactoryConfiguration>();
-    }
-
-    RuleTransformationPassConfiguration const& DefaultProfileGenerator::profilePassConfig() const
-    {
-        return configurationManager().get<RuleTransformationPassConfiguration>();
-    }
-
-    LlvmPassesConfiguration const& DefaultProfileGenerator::llvmConfig() const
-    {
-        return configurationManager().get<LlvmPassesConfiguration>();
-    }
-
-} // namespace quantum
-} // namespace microsoft
+}  // namespace quantum
+}  // namespace microsoft
